+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0] - UNRELEASED
+
+### Added
+
+- Skip only diagram wrapping with `--no-diagram-wrap`
+- Breaking of long elements over multiple lines in optional items `[]`
+
 ## [1.10.0] - 2020-09-16
 
 ### Added